feat: complete client page routing
