+++ src/client/index.js
@@ -3,6 +3,4 @@ import './styles/globalstyle.scss'
 
 import 'framework7-icons'
 
-import {} from './scripts/reportsList/pieChart'
-
-import {} from './scripts/route/router'
+import './scripts/router'
